add caching file to tikzcd svg

+++ mkdocs_mathenv_plugin/tikzcd.py
@@ -1,6 +1,7 @@
 from .tex import TeXWriter
 
 from hashlib import sha256
+from mkdocs.utils import log
 import os
 
 class TikZcdObject:
@@ -11,8 +12,21 @@ class TikZcdObject:
         self.options = options
         self.contents = contents
 
-    def write_to_svg(self) -> str:
+    def write_to_svg(self, cachefile: bool) -> str:
         filename = sha256(self.contents.encode()).hexdigest()
+
+        if cachefile == True:
+            try:
+                os.chdir("cache")
+            except OSError:
+                log.error("[mathenv] cache directory not found!")
+
+        if cachefile == True and os.path.exists(f"{filename}.svg"):
+            log.info("[mathenv] load from existed file...")
+            with open(f"{filename}.svg", "r", encoding="utf-8") as f:
+                svg_str = f.read(None)
+            os.chdir("..")
+            return svg_str
 
         writer = TeXWriter()
         writer.config.preamble = r'''
@@ -33,10 +47,13 @@ class TikZcdObject:
             svg_str = f.read(None)
 
         # clean up
-        try:
-            os.remove(filename + ".svg")
-        except FileNotFoundError:
-            pass
+        if cachefile == False :
+            try:
+                os.remove(filename + ".svg")
+            except FileNotFoundError:
+                pass
+
+        os.chdir("..")
 
         return svg_str
